tech(api): migrate repository to its bounded context
